requirements.txt
fastapi==0.115.12
openai_whisper==20240930
python_Levenshtein==0.27.1
